separate policy definition models from endpoints
