+++ catalystwan/endpoints/configuration/policy/definition/traffic_data.py
@@ -6,23 +6,18 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyDefinitionEndpoints
-from catalystwan.models.policy.definition.traffic_data import TrafficDataPolicy
+from catalystwan.models.policy.definition.traffic_data import (
+    TrafficDataPolicy,
+    TrafficDataPolicyEditPayload,
+    TrafficDataPolicyGetResponse,
+)
 from catalystwan.models.policy.policy_definition import (
     PolicyDefinitionEditResponse,
-    PolicyDefinitionGetResponse,
     PolicyDefinitionId,
     PolicyDefinitionInfo,
     PolicyDefinitionPreview,
 )
 from catalystwan.typed_list import DataSequence
-
-
-class TrafficDataPolicyEditPayload(TrafficDataPolicy, PolicyDefinitionId):
-    pass
-
-
-class TrafficDataPolicyGetResponse(TrafficDataPolicy, PolicyDefinitionGetResponse):
-    pass
 
 
 class ConfigurationPolicyDataDefinition(APIEndpoints, PolicyDefinitionEndpoints):
