+++ catalystwan/models/policy/definition/device_access.py
@@ -18,6 +18,8 @@ from catalystwan.models.policy.policy_definition import (
     Match,
     PolicyActionType,
     PolicyDefinitionBase,
+    PolicyDefinitionGetResponse,
+    PolicyDefinitionId,
     PolicyDefinitionSequenceBase,
     SourceDataPrefixListEntry,
     SourceIPEntry,
@@ -104,3 +106,11 @@ class DeviceAccessPolicy(DeviceAccessPolicyHeader, DefinitionWithSequencesCommon
             seq.match_device_access_protocol(port=device_access_protocol)
         self.add(seq)
         return seq
+
+
+class DeviceAccessPolicyEditPayload(DeviceAccessPolicy, PolicyDefinitionId):
+    pass
+
+
+class DeviceAccessPolicyGetResponse(DeviceAccessPolicy, PolicyDefinitionGetResponse):
+    pass
